Back-End/task_user/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-oauth-toolkit==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
jwcrypto==1.5.6
kombu==5.4.2
multidict==6.1.0
oauthlib==3.2.2
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.5
psycopg-binary==3.2.5
pycparser==2.22
python-dateutil==2.9.0.post0
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
djangorestframework-simplejwt==5.4.0
    # via -r Back-End/requirements.in
drf-yasg==1.21.8
    # via -r Back-End/requirements.in
gunicorn==22.0.0
minio==7.2.15
    # via
    #   -r requirements.in
    #   django-minio-storage
django-minio-storage==0.5.7
    # via -r requirements.in